requirements.txt
python==3.9.7
pytorch~=1.12.0
torch-geometric~=2.0.4
scikit-learn~=1.1.1
pyyaml~=6.0
trackml